Check for stale accessors to a collction embedded directly in a Mixed property
Change the index held by the collection object from ColKey to a structure
(ColIndex) )containing both the index of the column and a key generated for
that particular collection. The key value is stored alongside the ref and
compared with the key value found in index when trying to obtain the ref
for the collection.

+++ src/realm/collection_parent.cpp
@@ -316,10 +316,10 @@ int64_t CollectionParent::generate_key(size_t sz)
             key = int8_t(gen32());
         }
         else if (sz < 0x1000) {
-            key = int32_t(gen32());
+            key = int16_t(gen32());
         }
         else {
-            key = int64_t(gen32());
+            key = int32_t(gen32());
         }
     } while (key == 0);
